+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changes
 
+## Unreleased
+- Fixed Short URL with expire time
+
 ## 1.1.0
 - Page iteration
 - Add page iteration support for blacklist and contacts collection