+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changes
 
+## 1.1.1
+- Update http method for some endpoints
+
 ## 1.1.0
 - Page iteration
 - Add page iteration support for blacklist and contacts collection